Created a Static Page Module
Streamrr Static Module is to hold pages that are not going to be interacting with the database such as the home page, about page, terms of service, etc.

A new module will be created called the Authentication Module to handle the sign in and sign up pages for the app. This is where we will get the angular app to interact with some data.